feat: implement Expression

+++ src/main/java/view/ResultOutputView.java
@@ -1,7 +1,7 @@
 package view;
 
 public class ResultOutputView {
-    public void print(float result) {
+    public void print(double result) {
         System.out.println("결과 : " + result);
     }
 }
